Cleaned up DEBUG build warnings by commenting out unused variables in function declarations and bodies.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@1782 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/ptbl/xp/pf_Frag_TestRoutines.cpp
@@ -43,7 +43,7 @@ void pf_Frag::__dump(FILE * fp) const
 
 void pf_Frag_FmtMark::__dump(FILE * fp) const
 {
-	fprintf(fp,"        FmtMrkFragment %p api[%08lx]\n",this,m_indexAP);
+	fprintf(fp,"        FmtMrkFragment %p api[%08lx]\n",this,(long unsigned int)m_indexAP);
 }
 
 void pf_Frag_Strux_Block::__dump(FILE * fp) const
